Prepare for relase using OpenStack release process
xstatic-jquery-migrate has never been released using the OpenStack
release prociess with openstack/releases repository.
To sync the release history in openstack/releases and PyPI, we need to
release a new version, so the xstatic version is bumped.

README is also updated to satisfy what PyPI expects.
This is required to pass the validation in openstack/releases repo.

Change-Id: I642bc31efe778caec623cfb548cddcf57963e6f5