Add Javascript fork warning boilerplate to README
Following agreement at the Zed and Antelope PTGs, add a warning for
consumers to make them aware that we aren't the authors of the
Javascript files in this repository and aren't taking responsibility
for addressing security vulnerabilities in them. Adding it to the
README.rst ensures that it's prominent both when browsing the source
code as well as when looking at future versions of release pages on
PyPI, so that it should hopefully come to the attention of direct
users and redistributors like distro package maintainers alike.

Change-Id: I4cf50a2207abcdb8f050f5f2597ed6ebc635c13a